# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Greenhouse Backend API
  description: 농장 관리 시스템 Backend API
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/sales-slips:
    get:
      tags:
      - sales-controller
      operationId: getSalesSlips
      parameters:
      - name: partnerId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListSalesSlipResponse'
      deprecated: true
    post:
      tags:
      - sales-controller
      operationId: createSalesSlip
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SalesSlipCreateRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSalesSlipResponse'
...
  /api/orchid-groups:
    get:
      tags:
      - orchid-group-query-controller
      operationId: getOrchidGroups
      parameters:
      - name: houseId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: physicalBedId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: bedZoneId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListOrchidGroupResponse'
    post:
      tags:
      - orchid-group-command-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrchidGroupCreateRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrchidGroupResponse'
...
  /api/orchid-groups/{orchidGroupId}:
    get:
      tags:
      - orchid-group-query-controller
      operationId: getOrchidGroup
      parameters:
      - name: orchidGroupId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrchidGroupResponse'
    delete:
      tags:
      - orchid-group-command-controller
      operationId: delete_1
      parameters:
      - name: orchidGroupId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
    patch:
      tags:
      - orchid-group-command-controller
      operationId: update_4
      parameters:
      - name: orchidGroupId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrchidGroupUpdateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrchidGroupResponse'
...
  /api/work-operations/{workOperationId}/graph:
    get:
      tags:
      - work-operation-controller
      operationId: getGraph
      parameters:
      - name: workOperationId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: detail
        in: query
        required: false
        schema:
          type: string
          default: WORK
          enum:
          - WORK
          - MUTATION
          - LINEAGE
      - name: depth
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: maxNodes
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 120
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWorkOperationGraphResponse'
...
  /api/varieties/{varietyId}/orchid-groups:
    get:
      tags:
      - variety-controller
      operationId: getOrchidGroups_1
      parameters:
      - name: varietyId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListVarietyConnectedOrchidGroupResponse'
...
  /api/sales-slips/{salesSlipId}/print:
    get:
      tags:
      - print-controller
      operationId: getSalesSlipPrintData
      parameters:
      - name: salesSlipId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseSalesSlipResponse'
  /api/sales-slips/print:
    get:
      tags:
      - print-controller
      operationId: getPrintableSalesSlips
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseSalesSlipListItemResponse'
...
  /api/physical-beds:
    get:
      tags:
      - farm-structure-controller
      operationId: getPhysicalBeds
      parameters:
      - name: houseId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListPhysicalBedResponse'
  /api/physical-beds/{physicalBedId}:
    get:
      tags:
      - farm-structure-controller
      operationId: getPhysicalBed
      parameters:
      - name: physicalBedId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePhysicalBedResponse'
...
  /api/orchid-groups/{orchidGroupId}/lineage:
    get:
      tags:
      - orchid-group-query-controller
      operationId: getLineage
      parameters:
      - name: orchidGroupId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrchidGroupLineageResponse'
...
  /api/orchid-group-mutations:
    get:
      tags:
      - orchid-group-mutation-query-controller
      operationId: getMutations
      parameters:
      - name: orchidGroupId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: mutationType
        in: query
        required: false
        schema:
          type: string
          enum:
          - BASELINE_IMPORT
          - CREATE
          - UPDATE_DETAILS
          - MOVE
          - RESERVE
          - RELEASE_RESERVATION
          - CONSUME_RESERVATION
          - RESTORE_OUTBOUND
          - DISCARD
          - TRANSFORM
          - CORRECTION
          - RECONCILIATION
          - COMPENSATION
          - CANCEL_CREATION
          - DELETE
      - name: sourceDomain
        in: query
        required: false
        schema:
          type: string
          enum:
          - FARM
          - WORK
          - SALES
          - INBOUND
          - MIGRATION
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseOrchidGroupMutationResponse'
  /api/orchid-group-mutations/graph/{orchidGroupId}:
    get:
      tags:
      - orchid-group-mutation-query-controller
      operationId: getMutationGraph
      parameters:
      - name: orchidGroupId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: depth
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 2
      - name: maxNodes
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 120
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOrchidGroupMutationGraphResponse'
...
  /api/houses/{houseId}:
    get:
      tags:
      - farm-structure-controller
      operationId: getHouse
      parameters:
      - name: houseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseHouseResponse'
...
  /api/farm-status/orchid-groups:
    get:
      tags:
      - farm-status-controller
      operationId: getOrchidGroups_2
      parameters:
      - name: targetType
        in: query
        required: true
        schema:
          type: string
          enum:
          - HOUSE
          - PHYSICAL_BED
          - BED_ZONE
      - name: targetId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseFarmStatusOrchidGroupListResponse'
...
  /api/bed-zones:
    get:
      tags:
      - farm-structure-controller
      operationId: getBedZones
      parameters:
      - name: houseId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: physicalBedId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseListBedZoneResponse'
  /api/bed-zones/{bedZoneId}:
    get:
      tags:
      - farm-structure-controller
      operationId: getBedZone
      parameters:
      - name: bedZoneId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseBedZoneResponse'
...
components:
  schemas:
    SalesSlipCreateRequest:
      type: object
      properties:
        auctionShipmentId:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesSlipItemRequest'
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        partnerId:
          type: integer
          format: int64
        paymentMethod:
          type: string
          maxLength: 50
          minLength: 0
        paymentStatus:
          type: string
          maxLength: 50
          minLength: 0
        saleDate:
          type: string
          format: date
        salesStatus:
          type: string
          maxLength: 50
          minLength: 0
        salesType:
          type: string
          enum:
          - DIRECT
          - AUCTION
      required:
      - items
      - saleDate
    SalesSlipItemAllocationRequest:
      type: object
      properties:
        orchidGroupId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
          minimum: 1
      required:
      - orchidGroupId
      - quantity
    SalesSlipItemRequest:
      type: object
      properties:
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/SalesSlipItemAllocationRequest'
        genus:
          type: string
          maxLength: 100
          minLength: 0
        itemName:
          type: string
          maxLength: 150
          minLength: 0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
        spec:
          type: string
          maxLength: 100
          minLength: 0
        unitPrice:
          type: integer
          format: int32
          minimum: 0
      required:
      - allocations
      - itemName
      - quantity
      - unitPrice
    ApiResponseSalesSlipResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/SalesSlipResponse'
        message:
          type: string
...
    SalesSlipResponse:
      type: object
      properties:
        auctionMarket:
          type: string
        auctionShipmentId:
          type: integer
          format: int64
        availableActions:
          type: array
          items:
            type: string
            enum:
            - EDIT
            - COMPLETE
            - CANCEL
            - CONFIRM_PAYMENT
        expectedPaymentDate:
          type: string
          format: date
        id:
          type: integer
          format: int64
        items:
          type: array
          items:
            $ref: '#/components/schemas/SalesSlipItemResponse'
        memo:
          type: string
        paidAmount:
          type: integer
          format: int64
        partner:
          $ref: '#/components/schemas/BusinessPartnerResponse'
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        remainingAmount:
          type: integer
          format: int64
        saleDate:
          type: string
          format: date
        salesStatus:
          type: string
        salesType:
          type: string
          enum:
          - DIRECT
          - AUCTION
        slipNumber:
          type: string
        totalAmount:
          type: integer
          format: int32
...
    OrchidGroupResponse:
      type: object
      properties:
        ageYear:
          type: integer
          format: int32
        bedZoneId:
          type: integer
          format: int64
        bedZoneName:
          type: string
        endPosition:
          type: number
        genus:
          type: string
        houseId:
          type: integer
          format: int64
        houseNumber:
          type: integer
          format: int32
        id:
          type: integer
          format: int64
        memo:
          type: string
        physicalBedNumber:
          type: integer
          format: int32
        placementType:
          type: string
        potSize:
          type: string
        potSizeCode:
          type: string
          enum:
          - UNSPECIFIED
          - UNMAPPED
          - POT_2
          - POT_2_5
          - POT_3
          - POT_3_5
          - POT_4
          - POT_4_5
          - POT_5
          - POT_6
          - HANGING
          - ETC
        quantity:
          type: integer
          format: int32
        sortOrder:
          type: integer
          format: int32
        splitPlacementAllowed:
          type: boolean
        startPosition:
          type: number
        status:
          type: string
        trayCount:
          type: integer
          format: int32
        varietyColor:
          type: string
        varietyId:
          type: integer
          format: int64
        varietyName:
          type: string
...
    OrchidGroupCreateRequest:
      type: object
      properties:
        ageYear:
          type: integer
          format: int32
          minimum: 0
        bedZoneId:
          type: integer
          format: int64
        endPosition:
          type: number
          minimum: 0.0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        placementType:
          type: string
          maxLength: 50
          minLength: 0
        potSize:
          type: string
          maxLength: 50
          minLength: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
        splitPlacementAllowed:
          type: boolean
        startPosition:
          type: number
          minimum: 0.0
        status:
          type: string
          maxLength: 50
          minLength: 0
        trayCount:
          type: integer
          format: int32
          minimum: 0
        varietyId:
          type: integer
          format: int64
      required:
      - bedZoneId
      - quantity
      - status
      - varietyId
    ApiResponseOrchidGroupResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/OrchidGroupResponse'
        message:
          type: string
...
    ApiResponseVoid:
      type: object
      properties:
        data: {}
        message:
          type: string
...
    OrchidGroupUpdateRequest:
      type: object
      properties:
        ageYear:
          type: integer
          format: int32
          minimum: 0
        endPosition:
          type: number
          minimum: 0.0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        placementType:
          type: string
          maxLength: 50
          minLength: 0
        potSize:
          type: string
          maxLength: 50
          minLength: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
        splitPlacementAllowed:
          type: boolean
        startPosition:
          type: number
          minimum: 0.0
        status:
          type: string
          maxLength: 50
          minLength: 0
        trayCount:
          type: integer
          format: int32
          minimum: 0
        varietyId:
          type: integer
          format: int64
      required:
      - quantity
      - status
      - varietyId
...
    ApiResponseListOrchidGroupResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupResponse'
        message:
          type: string
...
    ApiResponseWorkOperationGraphResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/WorkOperationGraphResponse'
        message:
          type: string
...
    WorkOperationGraphResponse:
      type: object
      properties:
        depth:
          type: integer
          format: int32
        detail:
          type: string
          enum:
          - WORK
          - MUTATION
          - LINEAGE
        edges:
          type: array
          items:
            $ref: '#/components/schemas/WorkOperationGraphEdgeResponse'
        maxNodes:
          type: integer
          format: int32
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/WorkOperationGraphNodeResponse'
        rootWorkOperationId:
          type: integer
          format: int64
        truncated:
          type: boolean
...
    ApiResponseListVarietyConnectedOrchidGroupResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/VarietyConnectedOrchidGroupResponse'
        message:
          type: string
    VarietyConnectedOrchidGroupResponse:
      type: object
      properties:
        latestWorkDate:
          type: string
          format: date
        location:
          type: string
        orchidGroupId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
        status:
          type: string
...
    ApiResponseListSalesSlipResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SalesSlipResponse'
        message:
          type: string
    ApiResponsePageResponseSalesSlipListItemResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PageResponseSalesSlipListItemResponse'
        message:
          type: string
    PageResponseSalesSlipListItemResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SalesSlipListItemResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponseListPhysicalBedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalBedResponse'
        message:
          type: string
...
    PhysicalBedResponse:
      type: object
      properties:
        bedZones:
          type: array
          items:
            $ref: '#/components/schemas/BedZoneResponse'
        displayOrder:
          type: integer
          format: int32
        houseId:
          type: integer
          format: int64
        houseNumber:
          type: integer
          format: int32
        id:
          type: integer
          format: int64
        lengthCm:
          type: integer
          format: int32
        memo:
          type: string
        number:
          type: integer
          format: int32
        positionUnitCount:
          type: number
        positionUnitLabel:
          type: string
        supportIntervalCm:
          type: integer
          format: int32
        widthCm:
          type: integer
          format: int32
        wireCount:
          type: integer
          format: int32
    ApiResponsePhysicalBedResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PhysicalBedResponse'
        message:
          type: string
...
    ApiResponseOrchidGroupLineageResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/OrchidGroupLineageResponse'
        message:
          type: string
...
    OrchidGroupLineageResponse:
      type: object
      properties:
        orchidGroupId:
          type: integer
          format: int64
        results:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupLineageItemResponse'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupLineageItemResponse'
        transformations:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupLineageTransformationResponse'
...
    ApiResponsePageResponseOrchidGroupMutationResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PageResponseOrchidGroupMutationResponse'
        message:
          type: string
...
    PageResponseOrchidGroupMutationResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupMutationResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    ApiResponseOrchidGroupMutationGraphResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/OrchidGroupMutationGraphResponse'
        message:
          type: string
...
    OrchidGroupMutationGraphResponse:
      type: object
      properties:
        depth:
          type: integer
          format: int32
        edges:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupMutationGraphEdgeResponse'
        maxNodes:
          type: integer
          format: int32
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/OrchidGroupMutationGraphNodeResponse'
        rootOrchidGroupId:
          type: integer
          format: int64
        truncated:
          type: boolean
...
    HouseResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        memo:
          type: string
        name:
          type: string
        number:
          type: integer
          format: int32
        physicalBeds:
          type: array
          items:
            $ref: '#/components/schemas/PhysicalBedResponse'
    ApiResponseHouseResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/HouseResponse'
        message:
          type: string
...
    ApiResponseFarmStatusOrchidGroupListResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/FarmStatusOrchidGroupListResponse'
        message:
          type: string
...
    FarmStatusOrchidGroupListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FarmStatusOrchidGroupItemResponse'
        targetId:
          type: integer
          format: int64
        targetName:
          type: string
        targetType:
          type: string
          enum:
          - HOUSE
          - PHYSICAL_BED
          - BED_ZONE
...
    ApiResponseListBedZoneResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BedZoneResponse'
        message:
          type: string
    ApiResponseBedZoneResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BedZoneResponse'
        message:
          type: string
...